\name{foo.na} 
\docType{data}
\alias{foo.na} 
\title{Fictional incomplete time series with missing observations} 
\description{In the same format as \verb{foo} but with missing values. A data frame with fictional data. With the first column dates,  the second column a factor of what type of sample and the third column random numbers.}
\usage{foo.na} 
\format{A data.frame containing 1381 observations.} \keyword{datasets}