\name{foo} 
\docType{data}
\alias{foo} 
\title{Fictional incomplete time series} 
\description{A data frame with fictional data. With the first column dates,  the second column a factor of what type of sample and the third column random numbers.}
\usage{foo} 
\format{A data.frame containing 1381 observations.} \keyword{datasets}